scout-agent/requirements.txt
# Project dependencies
langchain==0.1.0
langchain-google-genai==0.0.11
langgraph==0.0.60
pydantic==2.5.0
PyGithub==2.1.1
python-dotenv==1.0.0
streamlit==1.28.0
sentence-transformers==2.2.2
scikit-learn==1.3.2
numpy==1.24.3
